Adapt rebased tests to main's Arc<ChoiceKind>/BigInt refactor
main reworked `ChoiceNode` to hold `Arc<ChoiceKind>` (built via
`ChoiceNode::new`) and made `IntegerChoice`/`ChoiceValue::Integer` use
`BigInt`. Update the tests added by this branch to match: construct nodes
with `ChoiceNode::new` and wrap integer values in `BigInt`.

(`git rebase -i` is unavailable in this environment, so this lands as a
follow-up rather than folded into the commits it adjusts.)

+++ tests/embedded/native/choices_tests.rs
@@ -791,9 +791,9 @@ fn choice_kind_unit_dispatches_to_each_sub_kind() {
     // replays) must forward to the matching sub-kind's `unit()` for every
     // variant.
     let ic = IntegerChoice {
-        min_value: 0,
-        max_value: 10,
-        shrink_towards: 0,
+        min_value: BigInt::from(0),
+        max_value: BigInt::from(10),
+        shrink_towards: BigInt::from(0),
     };
     assert_eq!(
         ChoiceKind::Integer(ic.clone()).unit(),
